requirements.txt
python-dotenv-0.21.1
flask-blueprint-1.3.0
flask_mysqldb-1.0.1
mysql_connector-2.2.9
flask-mysql-connector-1.1.0
Flask-JSON-0.3.5
datatypes-date-time-1.0.0a2
logging4-0.0.2